server/requirements.txt
aiosqlite==0.17.0
anyio==3.6.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
fastapi==0.85.0
greenlet==1.1.3
h11==0.14.0
idna==3.4
numpy==1.23.3
opencv-python==4.6.0.66
Pillow==9.2.0
pycparser==2.21
pydantic==1.10.2
PyJWT==2.5.0
python-decouple==3.6
python-multipart==0.0.5
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
torch==1.12.1
torchvision==0.13.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
